Move all client-side-only files to /imports/client/ so Meteor will do a quicker reload when modifying these files.
