+++ .gitignore
@@ -1,2 +1,3 @@
 .meteor
 node_modules
+npm-debug.log
