+++ app/client/main.js
@@ -7,18 +7,18 @@ import Router from 'react-router/BrowserRouter'
 import Match from 'react-router/Match'
 import Miss from 'react-router/Miss'
 
-import UserApp from '/imports/components/UserApp/UserApp.jsx'
-import Landing from '/imports/components/Landing/Landing.jsx'
-import Join from '/imports/components/Join/Join.jsx'
-import ContentPage from '/imports/components/ContentPage/ContentPage.jsx'
-import Login from '/imports/components/Login/Login.jsx'
-import CustomPage from '/imports/components/CustomPage/CustomPage.jsx'
-import Profile from '/imports/components/Profile/Profile.jsx'
-import LocationList from '/imports/containers/LocationList/LocationList.jsx'
-import LocationDetails from '/imports/containers/LocationDetails/LocationDetails.jsx'
-import ObjectDetails from '/imports/containers/ObjectDetails/ObjectDetails.jsx'
-import CommonBikeUI from '/imports/commonbike-ui.jsx'
-import NoMatch from '/imports/components/NoMatch/NoMatch.jsx'
+import UserApp from '/imports/client/components/UserApp/UserApp.jsx'
+import Landing from '/imports/client/components/Landing/Landing.jsx'
+import Join from '/imports/client/components/Join/Join.jsx'
+import ContentPage from '/imports/client/components/ContentPage/ContentPage.jsx'
+import Login from '/imports/client/components/Login/Login.jsx'
+import CustomPage from '/imports/client/components/CustomPage/CustomPage.jsx'
+import Profile from '/imports/client/components/Profile/Profile.jsx'
+import LocationList from '/imports/client/containers/LocationList/LocationList.jsx'
+import LocationDetails from '/imports/client/containers/LocationDetails/LocationDetails.jsx'
+import ObjectDetails from '/imports/client/containers/ObjectDetails/ObjectDetails.jsx'
+import CommonBikeUI from '/imports/client/commonbike-ui.jsx'
+import NoMatch from '/imports/client/components/NoMatch/NoMatch.jsx'
 
 Meteor.subscribe('locations')
