+++ app/imports/client/components/Avatar/Avatar.jsx
