+++ app/imports/client/components/CheckInCode/CheckInCode.jsx
