+++ app/imports/client/components/CommonBikeLogo/CommonBikeLogo.jsx
