+++ app/imports/client/components/ContentPage/ContentPage.jsx
