+++ app/imports/client/components/LocationList/LocationList.jsx
