+++ app/imports/client/components/SquareButton/SquareButton.jsx
