+++ app/imports/client/containers/LocationDetails/LocationDetails.jsx
