+++ app/imports/client/containers/LocationList/LocationList.jsx
