+++ app/imports/client/containers/ObjectDetails/ObjectDetails.jsx
